backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
boto3>=1.34.0
botocore>=1.34.0
faiss-cpu>=1.8.0
pypdf>=4.2.0
python-dotenv>=1.0.1
numpy>=1.26.0
bcrypt>=4.1.3
python-jose[cryptography]>=3.3.0
pydantic[email]

